ensure input is an IP address or resolves, at least by default
This helps address issue#4: we check if the given query string is
an IP address or if it resolves.  If not, we bail out, but the
user can choose to override this by specifying '-f', in which case
we will perform the whois lookups even if the given name doesn't
resolve (nor is an IP address).